requirements.txt
cleanco==2.3
langchain==0.3.12
langchain_community==0.3.12
langchain_core==0.3.25
langchain_openai==0.2.12
matplotlib==3.10.0
numpy==2.2.0
openai==1.57.4
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pandas==2.2.3
pdfservices_sdk==4.1.0
Pillow==11.0.0
polyfuzz==0.4.2
protobuf==5.29.1
pytesseract==0.3.13
python-dotenv==1.0.1
Requests==2.31.0
scikit_learn==1.6.0
selenium==4.27.1
stqdm==0.0.5
streamlit==1.40.2
